initial refactor of old codebase
