--- w2kplot/__init__.py
@@ -1,3 +1,0 @@
-from .w2kplot import Bands, FatBands, DensityOfStates, FermiSurface, WannierBands, ChargeDensity
-from .w2kplot import band_plot, fatband_plot, dos_plot, wannier_band_plot, charge_2d_plot
-import matplotlib.pyplot as plt
